+++ src/main/java/ro/sci/digitalCookBook/domain/AbstractModel.java
@@ -4,7 +4,7 @@ public abstract class AbstractModel{
 
     private int id;
 
-    public long getId(){ return id;}
+    public int getId(){ return id;}
 
     public void setId(int id) {this.id = id;}
